Strange warning: looks like the result of USHORT/UCHAR is being promoted to int and this causes a signed/unsigned wng.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@39034 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/cvt.cpp
@@ -434,7 +434,7 @@ static void validateLength(CharSet* toCharSet, SLONG toLength, const UCHAR* star
 		if (toCharSet->isMultiByte() &&
 			!(toCharSet->getFlags() & CHARSET_LEGACY_SEMANTICS) &&
 			toLength != 31 &&	// allow non CHARSET_LEGACY_SEMANTICS to be used as connection charset
-			toCharSet->length(toLength, start, false) > to_size / toCharSet->maxBytesPerChar())
+			toCharSet->length(toLength, start, false) > (ULONG) to_size / toCharSet->maxBytesPerChar())
 		{
 			err(Arg::Gds(isc_arith_except) << Arg::Gds(isc_transliteration_failed));
 		}
